+++ src/main/java/com/api/managepersonaddresses/models/PersonModel.java
@@ -23,7 +23,7 @@ public class PersonModel implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     
     @Column(name = "name", nullable = false, length = 45)
     private String name;
@@ -40,7 +40,7 @@ public class PersonModel implements Serializable {
     public PersonModel(String string, String string2) {
     }
     
-    public PersonModel(Integer id, String name, String birthDate, List<AddressModel> addresses) {
+    public PersonModel(Long id, String name, String birthDate, List<AddressModel> addresses) {
         this.id = id;
         this.name = name;
         this.birthDate = birthDate;
@@ -50,11 +50,11 @@ public class PersonModel implements Serializable {
     public PersonModel() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
     
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
